third_party/blink/web_tests/external/wpt/css/css-color/parsing/color-computed-color-mix-function-expected.txt

This is a testharness.js-based test.
Found 22 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] Property color value 'color-mix(in hsl, hsl(30deg 40% 80% / 25%) 0%, hsl(none none none / 0.5))'
  Colors do not match.\nActual:   color(srgb 0.48 0.4 0.32 / 0.5)\nExpected: color(srgb 0.88 0.8 0.72 / 0.5).\nError: assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 0.88 +/- 0.01, expected 0.88 but got 0.48
[FAIL] Property color value 'color-mix(in hsl, hsl(30deg 40% 80% / 25%) 0%, hsl(90deg none none / 0.5))'
  Colors do not match.\nActual:   color(srgb 0.4 0.48 0.32 / 0.5)\nExpected: color(srgb 0.8 0.88 0.72 / 0.5).\nError: assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 0.8 +/- 0.01, expected 0.8 but got 0.4
[FAIL] Property color value 'color-mix(in hwb, hwb(30deg 30% 40% / 25%) 0%, hwb(none none none / 0.5))'
  Colors do not match.\nActual:   color(srgb 0.8 0.475 0.15 / 0.5)\nExpected: color(srgb 0.6 0.45 0.3 / 0.5).\nError: assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 0.6 +/- 0.01, expected 0.6 but got 0.8
[FAIL] Property color value 'color-mix(in hwb, hwb(30deg 30% 40% / 25%) 0%, hwb(90deg none none / 0.5))'
  Colors do not match.\nActual:   color(srgb 0.475 0.8 0.15 / 0.5)\nExpected: color(srgb 0.45 0.6 0.3 / 0.5).\nError: assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 0.45 +/- 0.01, expected 0.45 but got 0.475
[FAIL] Property color value 'color-mix(in lch, lch(0.1 0.2 30deg / 25%) 0%, lch(none none none / 0.5))'
  Colors do not match.\nActual:   lch(0.05 0.1 30 / 0.5)\nExpected: lch(0.1 0.2 30 / 0.5).\nError: assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 0.1 +/- 0.01, expected 0.1 but got 0.05
[FAIL] Property color value 'color-mix(in lch, lch(0.1 0.2 30deg / 25%) 0%, lch(0.5 none none / 0.5))'
  Colors do not match.\nActual:   lch(0.5 0.1 30 / 0.5)\nExpected: lch(0.5 0.2 30 / 0.5).\nError: assert_array_approx_equals: Numeric parameters are approximately equal. property 1, expected 0.2 +/- 0.01, expected 0.2 but got 0.1
[FAIL] Property color value 'color-mix(in oklch, oklch(0.1 0.2 30deg / 25%) 0%, oklch(none none none / 0.5))'
  Colors do not match.\nActual:   oklch(0.05 0.1 30 / 0.5)\nExpected: oklch(0.1 0.2 30 / 0.5).\nError: assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 0.1 +/- 0.01, expected 0.1 but got 0.05
[FAIL] Property color value 'color-mix(in oklch, oklch(0.1 0.2 30deg / 25%) 0%, oklch(0.5 none none / 0.5))'
  Colors do not match.\nActual:   oklch(0.5 0.1 30 / 0.5)\nExpected: oklch(0.5 0.2 30 / 0.5).\nError: assert_array_approx_equals: Numeric parameters are approximately equal. property 1, expected 0.2 +/- 0.01, expected 0.2 but got 0.1
[FAIL] Property color value 'color-mix(in lab, lab(10 20 30 / 25%) 0%, lab(none none none / 0.5))'
  Colors do not match.\nActual:   lab(5 10 15 / 0.5)\nExpected: lab(10 20 30 / 0.5).\nError: assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 10 +/- 0.01, expected 10 but got 5
[FAIL] Property color value 'color-mix(in lab, lab(10 20 30 / 25%) 0%, lab(50 none none / 0.5))'
  Colors do not match.\nActual:   lab(50 10 15 / 0.5)\nExpected: lab(50 20 30 / 0.5).\nError: assert_array_approx_equals: Numeric parameters are approximately equal. property 1, expected 20 +/- 0.01, expected 20 but got 10
[FAIL] Property color value 'color-mix(in oklab, oklab(0.1 0.2 0.3 / 25%) 0%, oklab(none none none / 0.5))'
  Colors do not match.\nActual:   oklab(0.05 0.1 0.15 / 0.5)\nExpected: oklab(0.1 0.2 0.3 / 0.5).\nError: assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 0.1 +/- 0.01, expected 0.1 but got 0.05
[FAIL] Property color value 'color-mix(in oklab, oklab(0.1 0.2 0.3 / 25%) 0%, oklab(0.5 none none / 0.5))'
  Colors do not match.\nActual:   oklab(0.5 0.1 0.15 / 0.5)\nExpected: oklab(0.5 0.2 0.3 / 0.5).\nError: assert_array_approx_equals: Numeric parameters are approximately equal. property 1, expected 0.2 +/- 0.01, expected 0.2 but got 0.1
[FAIL] Property color value 'color-mix(in srgb, color(srgb .1 .2 .3 / 25%) 0%, color(srgb none none none / 50%))'
  Colors do not match.\nActual:   color(srgb 0.05 0.1 0.15 / 0.5)\nExpected: color(srgb 0.1 0.2 0.3 / 0.5).\nError: assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 0.1 +/- 0.01, expected 0.1 but got 0.05
[FAIL] Property color value 'color-mix(in srgb, color(srgb .1 .2 .3 / 25%) 0%, color(srgb 0.5 none none / 50%))'
  Colors do not match.\nActual:   color(srgb 0.5 0.1 0.15 / 0.5)\nExpected: color(srgb 0.5 0.2 0.3 / 0.5).\nError: assert_array_approx_equals: Numeric parameters are approximately equal. property 1, expected 0.2 +/- 0.01, expected 0.2 but got 0.1
[FAIL] Property color value 'color-mix(in srgb-linear, color(srgb-linear .1 .2 .3 / 25%) 0%, color(srgb-linear none none none / 50%))'
  Colors do not match.\nActual:   color(srgb-linear 0.05 0.1 0.15 / 0.5)\nExpected: color(srgb-linear 0.1 0.2 0.3 / 0.5).\nError: assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 0.1 +/- 0.01, expected 0.1 but got 0.05
[FAIL] Property color value 'color-mix(in srgb-linear, color(srgb-linear .1 .2 .3 / 25%) 0%, color(srgb-linear 0.5 none none / 50%))'
  Colors do not match.\nActual:   color(srgb-linear 0.5 0.1 0.15 / 0.5)\nExpected: color(srgb-linear 0.5 0.2 0.3 / 0.5).\nError: assert_array_approx_equals: Numeric parameters are approximately equal. property 1, expected 0.2 +/- 0.01, expected 0.2 but got 0.1
[FAIL] Property color value 'color-mix(in xyz, color(xyz .1 .2 .3 / 25%) 0%, color(xyz none none none / 50%))'
  Colors do not match.\nActual:   color(xyz-d65 0.05 0.1 0.15 / 0.5)\nExpected: color(xyz-d65 0.1 0.2 0.3 / 0.5).\nError: assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 0.1 +/- 0.01, expected 0.1 but got 0.05
[FAIL] Property color value 'color-mix(in xyz, color(xyz .1 .2 .3 / 25%) 0%, color(xyz 0.5 none none / 50%))'
  Colors do not match.\nActual:   color(xyz-d65 0.5 0.1 0.15 / 0.5)\nExpected: color(xyz-d65 0.5 0.2 0.3 / 0.5).\nError: assert_array_approx_equals: Numeric parameters are approximately equal. property 1, expected 0.2 +/- 0.01, expected 0.2 but got 0.1
[FAIL] Property color value 'color-mix(in xyz-d50, color(xyz-d50 .1 .2 .3 / 25%) 0%, color(xyz-d50 none none none / 50%))'
  Colors do not match.\nActual:   color(xyz-d50 0.05 0.1 0.15 / 0.5)\nExpected: color(xyz-d50 0.1 0.2 0.3 / 0.5).\nError: assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 0.1 +/- 0.01, expected 0.1 but got 0.05
[FAIL] Property color value 'color-mix(in xyz-d50, color(xyz-d50 .1 .2 .3 / 25%) 0%, color(xyz-d50 0.5 none none / 50%))'
  Colors do not match.\nActual:   color(xyz-d50 0.5 0.1 0.15 / 0.5)\nExpected: color(xyz-d50 0.5 0.2 0.3 / 0.5).\nError: assert_array_approx_equals: Numeric parameters are approximately equal. property 1, expected 0.2 +/- 0.01, expected 0.2 but got 0.1
[FAIL] Property color value 'color-mix(in xyz-d65, color(xyz-d65 .1 .2 .3 / 25%) 0%, color(xyz-d65 none none none / 50%))'
  Colors do not match.\nActual:   color(xyz-d65 0.05 0.1 0.15 / 0.5)\nExpected: color(xyz-d65 0.1 0.2 0.3 / 0.5).\nError: assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 0.1 +/- 0.01, expected 0.1 but got 0.05
[FAIL] Property color value 'color-mix(in xyz-d65, color(xyz-d65 .1 .2 .3 / 25%) 0%, color(xyz-d65 0.5 none none / 50%))'
  Colors do not match.\nActual:   color(xyz-d65 0.5 0.1 0.15 / 0.5)\nExpected: color(xyz-d65 0.5 0.2 0.3 / 0.5).\nError: assert_array_approx_equals: Numeric parameters are approximately equal. property 1, expected 0.2 +/- 0.01, expected 0.2 but got 0.1
Harness: the test ran to completion.
